+++ src/testSelfGit.test.ts
@@ -10,14 +10,15 @@ describe.skipIf(!!process.env.TEST_NO_SELF)("Git", () => {
 		"scans self",
 		async () => {
 			const files = gitFiles()
-			const r = await scan({ dirs: false, skipInternal: true, target: makeGit() })
+			const r = await scan({ dirs: false, target: makeGit() })
 			// this test uses sortFirstFolders implementation
 			// provided by https://jsr.io/@m234/path/0.1.4/sort-cmp.ts
 			// you can install this jsr package in your project
 			// for sorting - new Set(sorted) keeps sorting :),
 			// but your package and dependents should also declare
 			// @jsr:registry=https://npm.jsr.io in .npmrc or something.
-			expect(sortFirstFolders(r.paths.keys())).toMatchObject(sortFirstFolders(await files))
+			const keys = Array.from(r.paths.keys()).filter((p) => !p.startsWith(".bun/"))
+			expect(sortFirstFolders(keys)).toMatchObject(sortFirstFolders(await files))
 		},
 		{ timeout: 120e3 },
 	)